La til test for metrikkutsendelse ved print-klikk. Flyttet testen for autoutsendelse til samme fil.
