+++ src/Forside/Sammenligningspaneler/Sammenligningspaneler.tsx
@@ -11,11 +11,20 @@ import { sendIaTjenesteMetrikkMottatt } from '../../metrikker/iatjenester';
 interface Props {
     aggregertStatistikk: RestAggregertStatistikk;
     orgnr: string;
+    skalSendeMetrikkerAutomatisk?: boolean;
 }
 
-export const Sammenligningspaneler: FunctionComponent<Props> = ({ aggregertStatistikk, orgnr }) => {
+export const Sammenligningspaneler: FunctionComponent<Props> = ({
+    aggregertStatistikk,
+    orgnr,
+    skalSendeMetrikkerAutomatisk = true,
+}) => {
     useEffect(() => {
-        const timer = setTimeout(() => sendIaTjenesteMetrikkMottatt(orgnr), 5000);
+        const timer = setTimeout(() => {
+            if (skalSendeMetrikkerAutomatisk) {
+                sendIaTjenesteMetrikkMottatt(orgnr);
+            }
+        }, 5000);
         return () => clearTimeout(timer);
     }, [orgnr]);
